+++ src/main/java/com/haru/api/user/application/port/in/UserSignUpWorkflowUseCase.java
@@ -5,6 +5,8 @@ import com.haru.api.user.presentation.dto.UserResponseDTO;
 
 public interface UserSignUpWorkflowUseCase {
 
+    UserResponseDTO.User signUp(UserRequestDTO.SignUpRequest request, String token);
+
     UserResponseDTO.LoginResponse signUpAndLogin(UserRequestDTO.SignUpRequest request, String token);
 
 }
